% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_akl_council}
\alias{eolas_list_akl_council}
\title{List all Auckland Council datasets available in eolas}
\usage{
eolas_list_akl_council()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Auckland Council datasets available in eolas
}
