% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_bay_of_plenty}
\alias{eolas_list_bay_of_plenty}
\title{List all Bay of Plenty Councils datasets available in eolas}
\usage{
eolas_list_bay_of_plenty()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Bay of Plenty Councils datasets available in eolas
}
