% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_charities}
\alias{eolas_list_charities}
\title{List all Charities Services datasets available in eolas}
\usage{
eolas_list_charities()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Charities Services datasets available in eolas
}
