% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_colab_waikato}
\alias{eolas_list_colab_waikato}
\title{List all Co-Lab Waikato datasets available in eolas}
\usage{
eolas_list_colab_waikato()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Co-Lab Waikato datasets available in eolas
}
