% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_doc}
\alias{eolas_list_doc}
\title{List all DOC datasets available in eolas}
\usage{
eolas_list_doc()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all DOC datasets available in eolas
}
