% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_ecan_canterbury}
\alias{eolas_list_ecan_canterbury}
\title{List all ECan / Canterbury datasets available in eolas}
\usage{
eolas_list_ecan_canterbury()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all ECan / Canterbury datasets available in eolas
}
