% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_eeca}
\alias{eolas_list_eeca}
\title{List all EECA datasets available in eolas}
\usage{
eolas_list_eeca()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all EECA datasets available in eolas
}
