% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_geonet}
\alias{eolas_list_geonet}
\title{List all GeoNet datasets available in eolas}
\usage{
eolas_list_geonet()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all GeoNet datasets available in eolas
}
