% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_hawkes_bay}
\alias{eolas_list_hawkes_bay}
\title{List all Hawke's Bay Councils datasets available in eolas}
\usage{
eolas_list_hawkes_bay()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Hawke's Bay Councils datasets available in eolas
}
