% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_lris}
\alias{eolas_list_lris}
\title{List all Manaaki Whenua / LRIS datasets}
\usage{
eolas_list_lris()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Manaaki Whenua / LRIS datasets
}
