% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_napier_whanganui}
\alias{eolas_list_napier_whanganui}
\title{List all Napier + Whanganui datasets available in eolas}
\usage{
eolas_list_napier_whanganui()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Napier + Whanganui datasets available in eolas
}
