% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_otago}
\alias{eolas_list_otago}
\title{List all Otago Councils datasets available in eolas}
\usage{
eolas_list_otago()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Otago Councils datasets available in eolas
}
