% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_pharmac}
\alias{eolas_list_pharmac}
\title{List all PHARMAC datasets available in eolas}
\usage{
eolas_list_pharmac()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all PHARMAC datasets available in eolas
}
