% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_southland}
\alias{eolas_list_southland}
\title{List all Southland Councils datasets available in eolas}
\usage{
eolas_list_southland()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Southland Councils datasets available in eolas
}
