% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_statsnz}
\alias{eolas_list_statsnz}
\title{List all Stats NZ series}
\usage{
eolas_list_statsnz()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Stats NZ series
}
