% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_taranaki}
\alias{eolas_list_taranaki}
\title{List all Taranaki Councils datasets available in eolas}
\usage{
eolas_list_taranaki()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Taranaki Councils datasets available in eolas
}
