% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_top_of_south}
\alias{eolas_list_top_of_south}
\title{List all Gisborne / Top of South Councils datasets available in eolas}
\usage{
eolas_list_top_of_south()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Gisborne / Top of South Councils datasets available in eolas
}
