% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_wellington}
\alias{eolas_list_wellington}
\title{List all Wellington Region Councils datasets available in eolas}
\usage{
eolas_list_wellington()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all Wellington Region Councils datasets available in eolas
}
