% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{eolas_list_west_coast}
\alias{eolas_list_west_coast}
\title{List all West Coast (Te Tai o Poutini) datasets available in eolas}
\usage{
eolas_list_west_coast()
}
\value{
A data frame (tibble if available) of dataset metadata.
}
\description{
List all West Coast (Te Tai o Poutini) datasets available in eolas
}
